# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patoughi Driver API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/vehicles/{vehicle}": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Vehicle detail",
                "operationId": "vehicleDetail",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/VehicleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Update vehicle information",
                "operationId": "updateVehicle",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Vehicle update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/VehicleResource"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Delete vehicle status",
                "operationId": "deleteVehicleStatus",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "vehicle lists",
                "operationId": "vehicleLists",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/VehicleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Create Vehicle",
                "operationId": "createVehicle",
                "requestBody": {
                    "description": "Vehicle Store data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateVehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DriverVehicleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateVehicleRequest": {
                "required": [
                    "name",
                    "vehicle_type_id",
                    "fuel_type",
                    "plate_number",
                    "vin_number",
                    "insurance_policy_number",
                    "insurance_expiry_date",
                    "color",
                    "model_name",
                    "brand_name",
                    "year"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "vehicle_type_id": {
                        "type": "string"
                    },
                    "fuel_type": {
                        "type": "string",
                        "enum": [
                            "Gasoline",
                            "Diesel",
                            "CNG",
                            "Electric",
                            "Hybrid",
                            "Other"
                        ]
                    },
                    "plate_number": {
                        "type": "string"
                    },
                    "vin_number": {
                        "type": "string"
                    },
                    "insurance_policy_number": {
                        "type": "string"
                    },
                    "insurance_expiry_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "color": {
                        "type": "string"
                    },
                    "model_name": {
                        "type": "string"
                    },
                    "brand_name": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "VehicleUpdateRequest": {
                "title": "Vehicle Update Request",
                "description": "Request body for updating vehicle information",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "vehicle_type_id": {
                        "type": "string"
                    },
                    "fuel_type": {
                        "type": "string",
                        "enum": [
                            "Gasoline",
                            "Diesel",
                            "CNG",
                            "Electric",
                            "Hybrid",
                            "Other"
                        ]
                    },
                    "plate_number": {
                        "type": "string"
                    },
                    "vin_number": {
                        "type": "string"
                    },
                    "insurance_policy_number": {
                        "type": "string"
                    },
                    "insurance_expiry_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "color": {
                        "type": "string"
                    },
                    "model_name": {
                        "type": "string"
                    },
                    "brand_name": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "DriverVehicleResource": {
                "required": [
                    "id",
                    "name",
                    "vehicle_type",
                    "status",
                    "fuel_type",
                    "plate_number",
                    "vin_number",
                    "insurance_policy_number",
                    "insurance_expiry_date",
                    "color",
                    "model_name",
                    "brand_name",
                    "year"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeResource"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "maintenance",
                            "retired",
                            "inactive"
                        ]
                    },
                    "fuel_type": {
                        "type": "string",
                        "enum": [
                            "Gasoline",
                            "Diesel",
                            "CNG",
                            "Electric",
                            "Hybrid",
                            "Other"
                        ]
                    },
                    "plate_number": {
                        "type": "string"
                    },
                    "vin_number": {
                        "type": "string"
                    },
                    "insurance_policy_number": {
                        "type": "string"
                    },
                    "insurance_expiry_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "color": {
                        "type": "string"
                    },
                    "model_name": {
                        "type": "string"
                    },
                    "brand_name": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "VehicleResource": {
                "title": "Vehicle Resource",
                "description": "Vehicle resource response",
                "required": [
                    "id",
                    "name",
                    "vehicle_type",
                    "plate_number",
                    "color",
                    "status",
                    "fuel_type",
                    "vin_number",
                    "insurance_policy_number",
                    "insurance_expiry_date",
                    "current_location_lat",
                    "current_location_lng",
                    "model_name",
                    "brand_name",
                    "year",
                    "created_at",
                    "updated_at",
                    "is_allow_delete"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeResource"
                    },
                    "plate_number": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "maintenance",
                            "retired",
                            "inactive"
                        ]
                    },
                    "fuel_type": {
                        "type": "string",
                        "enum": [
                            "Gasoline",
                            "Diesel",
                            "CNG",
                            "Electric",
                            "Hybrid",
                            "Other"
                        ]
                    },
                    "vin_number": {
                        "type": "string"
                    },
                    "insurance_policy_number": {
                        "type": "string"
                    },
                    "insurance_expiry_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "current_location_lat": {
                        "type": "decimal"
                    },
                    "current_location_lng": {
                        "type": "decimal"
                    },
                    "model_name": {
                        "type": "string"
                    },
                    "brand_name": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "is_allow_delete": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ApiResponseErrorValidation": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorServer": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}